requirements.txt
pyqt6
easyocr
opencv-python
deepl
Pillow
requests
python-dotenv